refactor(base): simplify structure, add mjs extensions
