+++ packages/eslint-config-base/src/eslint.mjs
