+++ packages/eslint-config-base/src/index.mjs
@@ -3,7 +3,13 @@
  * @typedef {import('eslint').Linter.RulesRecord} RulesRecord
  */
 
-import { rules } from './rules';
+import { rules as eslint } from './eslint';
+import { rules as eslintPluginImport } from './eslint-plugin-import';
+
+const rules = {
+  ...eslint,
+  ...eslintPluginImport,
+};
 
 /**
  * Helper function for adding rules.
