--- packages/eslint-config-base/src/rules/index.js
@@ -1,11 +1,0 @@
-import { rules as eslint } from './eslint';
-import { rules as eslintPluginImport } from './eslint-plugin-import';
-
-const rules = {
-  ...eslint,
-  ...eslintPluginImport,
-};
-
-export {
-  rules,
-};
